+++ src/components/content/allergies/unread-allergies-notification.tsx
@@ -1,22 +1,10 @@
-import { useEffect, useState } from "react";
-import { useCTW } from "@/components/core/providers/use-ctw";
+import { useUserBuilderId } from "@/components/core/providers/user-builder-id";
 import { UnreadNotification } from "@/components/core/unread-notification";
 import { usePatientAllergies } from "@/fhir/allergies";
 
 export const UnreadAllergiesNotification = () => {
   const query = usePatientAllergies();
-  const { getRequestContext } = useCTW();
-
-  const [userBuilderId, setUserBuilderId] = useState("");
-
-  useEffect(() => {
-    async function load() {
-      const requestContext = await getRequestContext();
-      setUserBuilderId(requestContext.builderId);
-    }
-
-    void load();
-  }, [getRequestContext]);
+  const userBuilderId = useUserBuilderId();
 
   const unreadOutsideAllergies = query.data.filter(
     (allergy) => !allergy.isDismissed && !allergy.isRead && !allergy.ownedByBuilder(userBuilderId)
