+++ src/components/content/document/unread-documents-notification.tsx
@@ -1,22 +1,10 @@
-import { useEffect, useState } from "react";
-import { useCTW } from "@/components/core/providers/use-ctw";
+import { useUserBuilderId } from "@/components/core/providers/user-builder-id";
 import { UnreadNotification } from "@/components/core/unread-notification";
 import { usePatientDocuments } from "@/fhir/document";
 
 export const UnreadDocumentsNotification = () => {
   const query = usePatientDocuments();
-  const { getRequestContext } = useCTW();
-
-  const [userBuilderId, setUserBuilderId] = useState("");
-
-  useEffect(() => {
-    async function load() {
-      const requestContext = await getRequestContext();
-      setUserBuilderId(requestContext.builderId);
-    }
-
-    void load();
-  }, [getRequestContext]);
+  const userBuilderId = useUserBuilderId();
 
   const unreadOutsideDocuments = query.data.filter(
     (d) => !d.isDismissed && !d.isRead && !d.ownedByBuilder(userBuilderId)
