+++ src/components/content/immunizations/patient-immunizations.tsx
@@ -1,5 +1,4 @@
 import cx from "classnames";
-import { useEffect, useState } from "react";
 import { patientImmunizationsColumns } from "./helpers/columns";
 import { defaultImmunizationsFilters, immunizationsFilter } from "./helpers/filters";
 import { defaultImmunizationSort, immunizationSortOptions } from "./helpers/sort";
@@ -8,18 +7,18 @@ import { ResourceTable } from "../resource/resource-table";
 import { ResourceTableActions } from "../resource/resource-table-actions";
 import { entryFromArray } from "@/components/core/data-list";
 import { withErrorBoundary } from "@/components/core/error-boundary";
-import { useCTW } from "@/components/core/providers/use-ctw";
 import { usePatientImmunizations } from "@/fhir/immunizations";
 import { ImmunizationModel } from "@/fhir/models/immunization";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
 import { useFQSFeatureToggle } from "@/hooks/use-fqs-feature-toggle";
+import { useUserBuilderId } from "@/components/core/providers/user-builder-id";
 
 export type PatientImmunizationsProps = {
   className?: string;
 };
 
 function PatientImmunizationsComponent({ className }: PatientImmunizationsProps) {
-  const { getRequestContext } = useCTW();
+  const userBuilderId = useUserBuilderId();
   const { enabled } = useFQSFeatureToggle("immunizations");
   const patientImmunizationsQuery = usePatientImmunizations();
   const { data, setFilters, setSort } = useFilteredSortedData({
@@ -27,7 +26,6 @@ function PatientImmunizationsComponent({ className }: PatientImmunizationsProps)
     defaultSort: defaultImmunizationSort,
     records: patientImmunizationsQuery.data,
   });
-  const [userBuilderId, setUserBuilderId] = useState("");
 
   const openDetails = useResourceDetailsDrawer({
     header: (m) => m.description,
@@ -35,15 +33,6 @@ function PatientImmunizationsComponent({ className }: PatientImmunizationsProps)
     getSourceDocument: true,
     enableFQS: enabled,
   });
-
-  useEffect(() => {
-    async function load() {
-      const requestContext = await getRequestContext();
-      setUserBuilderId(requestContext.builderId);
-    }
-
-    void load();
-  }, [getRequestContext]);
 
   return (
     <div
