+++ src/components/content/immunizations/unread-immunizations-notification.tsx
@@ -1,21 +1,9 @@
-import { useEffect, useState } from "react";
-import { useCTW } from "@/components/core/providers/use-ctw";
+import { useUserBuilderId } from "@/components/core/providers/user-builder-id";
 import { usePatientImmunizations } from "@/fhir/immunizations";
 
 export const UnreadImmunizationsNotification = () => {
   const query = usePatientImmunizations();
-  const { getRequestContext } = useCTW();
-
-  const [userBuilderId, setUserBuilderId] = useState("");
-
-  useEffect(() => {
-    async function load() {
-      const requestContext = await getRequestContext();
-      setUserBuilderId(requestContext.builderId);
-    }
-
-    void load();
-  }, [getRequestContext]);
+  const userBuilderId = useUserBuilderId();
 
   const unreadOutsideImmunizations = query.data.filter(
     (imm) => !imm.isDismissed && !imm.isRead && !imm.ownedByBuilder(userBuilderId)
